+++ interview/tests.py
@@ -1,21 +1,32 @@
-from django.conf import settings
-from django.core import mail
-from django.test import TestCase, RequestFactory
 import datetime
 import hashlib
 
-from django.urls import reverse
-
-from interview.factory import ProcessFactory, InterviewFactory, CandidateFactory
-from interview.models import Process, Document, Interview, Candidate
-
+import dateutil.relativedelta
 import pytz
+from django.conf import settings
+from django.core import mail
+from django.test import TestCase, RequestFactory
+from django.urls import reverse
+from django.utils.text import slugify
+from django.utils.translation import ugettext_lazy as _
+from factory.faker import faker
 
+from interview import views
+from interview.factory import (
+    ProcessFactory,
+    InterviewFactory,
+    CandidateFactory,
+    ContractTypeFactory,
+    SourcesFactory,
+    OfferFactory,
+    InterviewKindFactory,
+)
+from interview.models import Process, Document, Interview
 from interview.views import process, minute_edit, minute, interview, close_process, reopen_process
 from ref.factory import SubsidiaryFactory, ConsultantFactory
 from ref.models import Consultant
 
-from django.conf import settings
+from django.utils.translation import activate
 
 
 class InterviewTestCase(TestCase):
@@ -330,3 +341,594 @@ class AnonymizesCanditateTestCase(TestCase):
         self.assertEqual(other_candidate2.anonymized_email(), email_hash.hexdigest())
         previous_candidate_anoymized = other_candidate2.find_duplicates()
         self.assertEqual(1, previous_candidate_anoymized.count())
+
+
+class HomeViewTestCase(TestCase):
+
+    # subsidiary_processes_table.data => class TableQuerysetData(TableData) in
+    # https://django-tables2.readthedocs.io/en/stable/_modules/django_tables2/data.html
+
+    def setUp(self) -> None:
+        self.url = reverse(views.dashboard)
+        self.assertEqual(self.url, "/")
+
+    def test_dashboard_not_logged_in(self):
+        response = self.client.get(self.url)
+        self.assertRedirects(response, f"/admin/login/?next={self.url}")
+
+    def test_dashboard_logged_in(self):
+        # create a consultant
+        subsidiary = SubsidiaryFactory()
+        consultant = ConsultantFactory(company=subsidiary)
+        user = consultant.user
+
+        # log user in
+        self.client.force_login(user=user)
+
+        # access dashboard
+        response = self.client.get(self.url)
+
+        # assert we were allowed to access dashboard
+        self.assertEqual(response.status_code, 200)
+        # assert the right template was called
+        self.assertTemplateUsed(response, template_name="interview/dashboard.html")
+        # assert the header was rendered
+        self.assertTemplateUsed(response, template_name="interview/base.html")
+        # assert some tables were rendered inside the dashboard
+        self.assertTemplateUsed(response, template_name="interview/_tables.html")
+
+    def test_dashboard_no_data(self):
+        # create a consultant
+        subsidiary = SubsidiaryFactory()
+        consultant = ConsultantFactory(company=subsidiary)
+        user = consultant.user
+
+        # log user in
+        self.client.force_login(user=user)
+
+        # access dashboard
+        response = self.client.get(self.url)
+        # retrieve the three tables displayed
+        actions_needed_processes_table = response.context["actions_needed_processes_table"]
+        related_processes_table = response.context["related_processes_table"]
+        subsidiary_processes_table = response.context["subsidiary_processes_table"]
+
+        # assert all three tables are empty
+        processes = actions_needed_processes_table.data
+        self.assertEqual(len(processes), 0)
+
+        processes = related_processes_table.data
+        self.assertEqual(len(processes), 0)
+
+        processes = subsidiary_processes_table.data
+        self.assertEqual(len(processes), 0)
+
+    def test_dashboard_with_data(self):
+        # create a consultant
+        subsidiary = SubsidiaryFactory()
+        consultant = ConsultantFactory(subsidiary=subsidiary.id)
+        user = consultant.user
+
+        # create 1 process in actions_needed_processes_table & in related_process_table
+        p1 = ProcessFactory()
+        itw1 = InterviewFactory(process=p1)
+        itw1.interviewers.add(consultant)
+        itw1.save()
+
+        # create 1 process only appearing in related_process_table
+        p2 = ProcessFactory()
+        itw2 = InterviewFactory(process=p2)
+        itw2.interviewers.add(consultant)
+        itw2.save()
+        p2.end_date = datetime.datetime.now()
+        p2.state = Process.OTHER
+        p2.save()
+
+        # create a process that will only appear in subsidiary_process_table
+        ProcessFactory(subsidiary=subsidiary)
+
+        # log user in
+        self.client.force_login(user=user)
+
+        # access dashboard
+        response = self.client.get(self.url)
+
+        # retrieve the three tables displayed
+        actions_needed_processes_table = response.context["actions_needed_processes_table"]
+        related_processes_table = response.context["related_processes_table"]
+        subsidiary_processes_table = response.context["subsidiary_processes_table"]
+
+        # assert all three tables are correctly displayed
+        processes_actions_needed = actions_needed_processes_table.data
+        self.assertEqual(len(processes_actions_needed), 1)
+        self.assertEqual(p1.id, processes_actions_needed[0].id)
+
+        related_processes = related_processes_table.data
+        self.assertEqual(len(related_processes), 2)
+        ids = []
+        for tmp in related_processes:
+            ids.append(tmp.id)
+        self.assertTrue(p1.id in ids)
+        self.assertTrue(p2.id in ids)
+
+        subsidiary_processes = subsidiary_processes_table.data
+        self.assertEqual(len(subsidiary_processes), 1)
+        self.assertNotEquals(p1.id, subsidiary_processes[0].id)
+        self.assertNotEquals(p2.id, subsidiary_processes[0].id)
+
+
+class ProcessCreationViewTestCase(TestCase):
+    def setUp(self):
+        self.url = reverse(views.new_candidate)
+        self.assertEqual(self.url, "/candidate/")
+
+        # create a consultant
+        self.subsidiary = SubsidiaryFactory()
+        self.consultant = ConsultantFactory(company=self.subsidiary)
+        self.user = self.consultant.user
+
+        self.fake = faker.Faker()
+
+        self.tz = pytz.timezone("Europe/Paris")
+
+    def test_process_creation_not_logged_in(self):
+        response = self.client.get(self.url)
+        self.assertRedirects(response, f"/admin/login/?next={self.url}")
+
+    def test_process_creation_logged_in(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # access process creation view
+        response = self.client.get(self.url)
+
+        # assert we were allowed to access this view
+        self.assertEqual(response.status_code, 200)
+        # assert the right template was called
+        self.assertTemplateUsed(response, template_name="interview/new_candidate.html")
+        # assert the header was rendered
+        self.assertTemplateUsed(response, template_name="interview/base.html")
+
+    def test_assert_all_needed_forms_are_displayed(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # access dashboard
+        response = self.client.get(self.url)
+
+        context = response.context
+        candidate_form = context["candidate_form"]
+        process_form = context["process_form"]
+        source_form = context["source_form"]
+        offer_form = context["offer_form"]
+        interviewers_form = context["interviewers_form"]
+        duplicate_processes = context["duplicates"]
+        candidate = context["candidate"]
+        self.assertIsNotNone(candidate_form)
+        self.assertIsNotNone(process_form)
+        self.assertIsNotNone(source_form)
+        self.assertIsNotNone(offer_form)
+        self.assertIsNotNone(interviewers_form)
+        self.assertIsNone(duplicate_processes)
+        self.assertIsNone(candidate)
+
+    def test_form_submit_with_least_data(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        candidate_name = self.fake.name()
+        response = self.client.post(
+            path=reverse(views.new_candidate),
+            data={
+                "name": candidate_name,
+                "subsidiary": self.subsidiary.id,
+                "summit": "Enregistrer",  # means having clicked on "Save" in the page
+                "new-candidate": True,  # bypass checks for already existing candidate
+            },
+            follow=True,
+        )
+
+        p = Process.objects.filter(candidate__name=candidate_name).first()
+        self.assertIsNotNone(p)
+        self.assertRedirects(
+            response,
+            p.get_absolute_url(),
+        )
+
+    def test_form_submit_with_full_data(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        candidate_name = self.fake.name()
+        candidate_mail = f"{slugify(candidate_name)}@mail.com"
+        candidate_phone = f"+336{self.fake.random_number(fix_len=True, digits=8)}"
+        process_sub = self.subsidiary
+        contract_type = ContractTypeFactory()
+        expected_salary = 40
+        contract_length = 24
+        contract_start_date = datetime.date.today() + dateutil.relativedelta.relativedelta(months=1)
+        source = SourcesFactory()
+        offer = OfferFactory()
+        complementary_informations = "This is a very useful comment"
+        interviewers = self.consultant
+        interview_kind = InterviewKindFactory()
+
+        response = self.client.post(
+            path=reverse(views.new_candidate),
+            data={
+                "name": candidate_name,
+                "email": candidate_mail,
+                "phone": candidate_phone,
+                "subsidiary": process_sub.id,
+                "contract_type": contract_type.id,
+                "salary_expectation": expected_salary,
+                "contract_duration": contract_length,
+                "contract_start_date": contract_start_date,
+                "sources": source.id,
+                "offer": offer.id,
+                "other_informations": complementary_informations,
+                "interviewers-kind_of_interview": interview_kind.id,
+                "interviewers-interviewers": interviewers.id,
+                "summit": "Enregistrer",  # means having clicked on "Save" in the page
+                "new-candidate": True,  # bypass checks for already existing candidate
+            },
+            follow=True,
+        )
+
+        p = Process.objects.filter(candidate__name=candidate_name).first()
+        self.assertIsNotNone(p)
+        self.assertRedirects(
+            response,
+            p.get_absolute_url(),
+        )
+
+    def test_correct_display_of_existing_candidate(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # create a new candidate
+        candidate_name = self.fake.name()
+        process_sub = self.subsidiary
+        offer = OfferFactory()
+        response = self.client.post(
+            path=reverse(views.new_candidate),
+            data={
+                "name": candidate_name,
+                "subsidiary": process_sub.id,
+                "offer": offer.id,
+                "summit": "Enregistrer",  # means having clicked on "Save" in the page
+                "new-candidate": True,  # bypass checks for already existing candidate
+            },
+            follow=True,
+        )
+
+        # assert the candidate was correctly created
+        p = Process.objects.filter(candidate__name=candidate_name).first()
+        self.assertIsNotNone(p)
+        self.assertRedirects(
+            response,
+            p.get_absolute_url(),
+        )
+
+        # same call
+        response = self.client.post(
+            path=reverse(views.new_candidate),
+            data={
+                "name": candidate_name,
+                "subsidiary": process_sub.id,
+                "offer": offer.id,
+                "summit": "Enregistrer",  # means having clicked on "Save" in the page
+            },
+            follow=True,
+        )
+
+        self.assertEqual(response.status_code, 200)
+        self.assertIsNotNone(response.context["duplicates"])
+
+    def test_form_submit_reusing_candidate(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # create a new candidate
+        candidate_name = self.fake.name()
+        process_sub = self.subsidiary
+        offer = OfferFactory()
+        response = self.client.post(
+            path=reverse(views.new_candidate),
+            data={
+                "name": candidate_name,
+                "subsidiary": process_sub.id,
+                "offer": offer.id,
+                "summit": "Enregistrer",  # means having clicked on "Save" in the page
+                "new-candidate": True,  # bypass checks for already existing candidate
+            },
+            follow=True,
+        )
+
+        # assert the candidate was correctly created
+        p = Process.objects.filter(candidate__name=candidate_name).first()
+        self.assertIsNotNone(p)
+        self.assertRedirects(response, p.get_absolute_url())
+
+        new_offer = OfferFactory()
+
+        response = self.client.post(
+            path=reverse(views.reuse_candidate, kwargs={"candidate_id": p.candidate.id}),
+            data={
+                "name": candidate_name,
+                "subsidiary": process_sub.id,
+                "offer": new_offer.id,
+                "email": "",
+                "phone": "",
+                "summit": "Enregistrer",
+            },
+            follow=True,
+        )
+
+        process_for_candidate = Process.objects.filter(candidate__name=candidate_name).count()
+        self.assertEqual(process_for_candidate, 2)
+        p = Process.objects.filter(candidate__name=candidate_name).last()
+        self.assertIsNotNone(p)
+        self.assertRedirects(response, p.get_absolute_url())
+
+
+class ProcessDetailsViewTestCase(TestCase):
+    def setUp(self):
+        # create a process
+        self.subsidiary = SubsidiaryFactory()
+        self.consultant = ConsultantFactory(company=self.subsidiary)
+        self.user = self.consultant.user
+        self.process = ProcessFactory(
+            subsidiary=self.subsidiary,
+            contract_type=ContractTypeFactory(),
+            offer=OfferFactory(),
+            sources=SourcesFactory(),
+        )
+        self.candidate = self.process.candidate
+        self.url = reverse(
+            views.process, kwargs={"process_id": self.process.id, "slug_info": self.process.candidate.name_slug}
+        )
+        self.assertEqual(self.url, f"/process/{self.process.id}{self.process.candidate.name_slug}/")
+
+        activate("en")
+
+    def test_process_details_not_logged_in(self):
+        response = self.client.get(self.url)
+        self.assertRedirects(response, f"/admin/login/?next={self.url}")
+
+    def test_process_details_logged_in(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # access process details
+        response = self.client.get(self.url)
+
+        # assert we were allowed to access our process details view
+        self.assertEqual(response.status_code, 200)
+        # assert the right template was called
+        self.assertTemplateUsed(response, template_name="interview/process_detail.html")
+        # assert the header was rendered
+        self.assertTemplateUsed(response, template_name="interview/base.html")
+
+    def test_process_details_info_display(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # access process details
+        response = self.client.get(self.url)
+
+        self.assertEqual(response.status_code, 200)
+
+        self.assertContains(
+            response,
+            reverse(views.edit_candidate, kwargs={"process_id": self.process.id}),
+        )
+
+        self.assertContains(response, self.candidate.name)
+        self.assertContains(response, self.subsidiary)
+
+        # Informations de contact
+        self.assertContains(response, self.candidate.email)
+        self.assertContains(response, self.candidate.phone)
+
+        # Informations sur le contrat
+        self.assertContains(response, self.process.sources)
+        self.assertContains(response, self.process.offer.name)
+        self.assertContains(response, "Documents")
+
+        self.assertContains(response, self.process.contract_type)
+        self.assertContains(response, self.process.contract_duration)
+
+        self.assertContains(response, f"{self.process.salary_expectation} k€")
+
+        self.assertContains(response, self.process.other_informations)
+
+        # assert that interview tables is empty
+        interviews = response.context["interviews_for_process_table"].data
+        self.assertFalse(interviews)
+
+        # assert that add itw button exists
+        self.assertContains(response, "Add an interview")
+
+        # close the process
+        response = self.client.post(
+            path=reverse(views.close_process, kwargs={"process_id": self.process.id}),
+            data={"state": "NG", "closed_comment": "closed", "summit": "Terminer+le+processus"},
+            follow=True,
+        )
+
+        # update process values after state update
+        self.process = Process.objects.get(id=self.process.id)
+
+        self.assertRedirects(response, self.process.get_absolute_url())
+        self.assertContains(response, self.process.get_state_display())
+        self.assertContains(response, self.process.closed_comment)
+
+    def test_process_details_interviews_table(self):
+        # new process as it was closed by last test
+        self.setUp()
+
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # add an interview
+        itw1 = InterviewFactory(process=self.process, kind_of_interview=InterviewKindFactory())
+        itw1.interviewers.add(self.consultant)  # needed to be able to render the process page
+        itw1.save()
+
+        # access process details
+        response = self.client.get(self.url)
+
+        interviews = response.context["interviews_for_process_table"].data
+        self.assertEqual(len(interviews), 1)
+        self.assertEqual(
+            str(interviews[0]), str(itw1)
+        )  # string representation should be enough to differentiate each itw
+
+        # add an interview
+        itw2 = InterviewFactory(process=self.process, kind_of_interview=InterviewKindFactory())
+        itw2.interviewers.add(self.consultant)  # needed to be able to render the process page
+        itw2.save()
+
+        # access process details
+        response = self.client.get(self.url)
+
+        interviews = response.context["interviews_for_process_table"].data
+        self.assertEqual(len(interviews), 2)
+        self.assertEqual(
+            str(interviews[1]), str(itw2)
+        )  # string representation should be enough to differentiate each itw
+
+
+class InterviewMinuteViewTestCase(TestCase):
+    def setUp(self):
+        self.subsidiary = SubsidiaryFactory()
+        self.consultant = ConsultantFactory(company=self.subsidiary)
+        self.user = self.consultant.user
+        self.process = ProcessFactory(
+            subsidiary=self.subsidiary,
+            contract_type=ContractTypeFactory(),
+            offer=OfferFactory(),
+            sources=SourcesFactory(),
+        )
+        self.candidate = self.process.candidate
+
+        self.itw = InterviewFactory(process=self.process, kind_of_interview=InterviewKindFactory())
+        self.itw.interviewers.add(self.consultant)
+        self.itw.save()
+
+        self.url_minute = reverse(
+            views.minute, kwargs={"interview_id": self.itw.id, "slug_info": self.process.candidate.name_slug}
+        )
+        self.assertEqual(
+            self.url_minute,
+            f"/interview/{self.itw.id}{self.process.candidate.name_slug}/minute/",
+        )
+
+        self.url_edit = reverse(views.minute_edit, kwargs={"interview_id": self.itw.id})
+        self.assertEqual(self.url_edit, f"/interview/{self.itw.id}/minute/edit/")
+
+        activate("en")
+
+    def test_interview_minute_not_logged_in(self):
+        response = self.client.get(self.url_minute)
+        self.assertRedirects(response, f"/admin/login/?next={self.url_minute}")
+
+    def test_interview_minute_edit_not_logged_in(self):
+        response = self.client.get(self.url_edit)
+        self.assertRedirects(response, f"/admin/login/?next={self.url_edit}")
+
+    def test_interview_minute_logged_in(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # access process creation view
+        response = self.client.get(self.url_minute)
+
+        # assert we were allowed to access this view
+        self.assertEqual(response.status_code, 200)
+        # assert the right template was called
+        self.assertTemplateUsed(response, template_name="interview/interview_minute.html")
+        # assert the header was rendered
+        self.assertTemplateUsed(response, template_name="interview/base.html")
+
+    def test_interview_minute_edit_logged_in(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # access process creation view
+        response = self.client.get(self.url_edit)
+
+        # assert we were allowed to access this view
+        self.assertEqual(response.status_code, 200)
+        # assert the right template was called
+        self.assertTemplateUsed(response, template_name="interview/interview_minute_form.html")
+        # assert the header was rendered
+        self.assertTemplateUsed(response, template_name="interview/base.html")
+
+    def test_interview_minute_display(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # access process creation view
+        response = self.client.get(self.url_minute)
+
+        self.assertEqual(response.status_code, 200)
+
+        self.assertContains(response, self.subsidiary)
+        self.assertContains(response, reverse(views.edit_candidate, kwargs={"process_id": self.process.id}))
+        self.assertContains(response, self.candidate.name)
+
+        self.assertContains(
+            response,
+            self.process.get_absolute_url(),
+        )
+
+        # Interviewers (only one tho)
+        self.assertContains(response, self.itw.interviewers.first())
+
+        # CR
+        self.assertContains(response, self.itw.kind_of_interview)
+        self.assertContains(response, self.itw.minute)
+
+        # Suggested interviewer
+        self.assertContains(response, self.itw.suggested_interviewer)
+
+        # Next itw goel
+        self.assertContains(response, self.itw.next_interview_goal)
+
+    def test_interview_minute_edit_display(self):
+        # log user in
+        self.client.force_login(user=self.user)
+
+        # access process creation view
+        response = self.client.get(self.url_edit)
+
+        self.assertEqual(response.status_code, 200)
+
+        #  Compte rendu d'entretien pour <assert 3> pour <assert 2> pour la filiale <assert 1>
+        self.assertContains(response, self.subsidiary)
+        self.assertContains(response, reverse(views.edit_candidate, kwargs={"process_id": self.process.id}))
+        self.assertContains(response, self.candidate.name)
+
+        self.assertContains(
+            response,
+            self.process.get_absolute_url(),
+        )
+
+        # CR
+        self.assertContains(response, self.itw.minute)
+
+        # Suggested interviewer
+        self.assertContains(response, self.itw.suggested_interviewer)
+
+        # Suggested interviewer
+        self.assertContains(response, self.itw.suggested_interviewer)
+
+        # Type d'entretien
+        self.assertContains(response, self.itw.kind_of_interview)
+
+        # assert buttons are here
+        self.assertContains(response, "NO")
+        self.assertContains(response, "DRAFT")
+        self.assertContains(response, "GO")