+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - More refined factories
 - New command: create_dev_dataset to generate a basic dev dataset
 - Add process creator, creation date and linkedin url on process page
+- Add view tests
 - Add interview planned date in process creation
 
 # v1.15.0 (2022-10-20)